Allow theming of framework UI elements.
As the theme engine works by replacing resources while loading the APK,
framework UI couldn't be themed so far. Workaround that by manually
creating a com.android.systemui package context and using that for
showing the UI. This attempt is a bit hacky, but should be safe as
a) systemui should always be present
b) if it isn't, we fall back to no theming

Change-Id: I644e2e8d7d4abf07bae639b606197d691f16adac

Applying theming to more framework items.

All dialogs and notifications should be covered now.

Change-Id: If93f81991361165a00b5b1b03ffc323e3ab42c0b