Add Deskdock support to WiredAccessoryObserver
After discussion on XDA it appears that Samsung Desk Docks send a signal of 1.
Car Docs send a signal of 2

This change should work with Desk Docks now, routing the audio when the option is checked in the Dock Menu.

I have tested that no regression has occured - Car Dock audio is working.

Change-Id: Ie89a9a1c9738676188b53f9d6b22db92b6894b8c