SystemUI: Avoid "phantom" status bar collapsing
There is one specific case where the status bar may collapse
for no aparent reason: When the last notification is removed.

If Quick Settings panel is showing while such a notification
is removed, the status bar WILL collapse -- and the user may
be left wondering why, because the notification that was just
removed isn't visible in that case.

A way to reproduce this behaviour follows:
  1) Settings -> System -> Quick Settings Panel -> uncheck
     "Auto close panel"
  2) Dismiss all notifications, even the sticky ones.
  3) Open the Quick Settings Panel and tap the apropriate
     tile to light the Torch
  4) Tap the tile again, to extinguish the Torch.

As you can see, the Quick Settings Panel collapses. This
happens because Torch puts up a notification, when lit and
removes it, just after it's turned off.

This will happen to whatever notification that hides itself,
e.g. a GMail notification, which vanishes when you read the
e-mail on another device. This too will trigger this issue.

-----

This patch prevents the bar from collapsing, as consequence
of removing the last notification, when Quick Settings is
showing.

Change-Id: I8c7c015c836810a465268b25b90c9166c4485a9f